LayoutTests/fast/css3-text/css3-text-decoration/getComputedStyle/getComputedStyle-text-decoration-shorthand-expected.txt

Test to make sure -webkit-text-decoration property is backwards compatible with CSS 2.1 and CSS 3 shorthand.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial values:
PASS e.style.getPropertyCSSValue('-webkit-text-decoration') is null
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration').cssText is 'none solid rgb(0, 0, 0)'
PASS checkComputedStyleValue() is true

CSS2.1 backwards compatibility ("-webkit-text-decoration: underline overline line-through blink;"):
PASS e.style.getPropertyCSSValue('-webkit-text-decoration-line').cssText is 'underline overline line-through blink'
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration-line').cssText is 'underline overline line-through'
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration').cssText is 'underline overline line-through solid rgb(0, 0, 0)'
PASS checkComputedStyleValue() is true

CSS3 Shorthand ("-webkit-text-decoration: underline overline line-through blink dashed red;"):
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration').cssText is 'underline overline line-through dashed rgb(255, 0, 0)'
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration-line').cssText is 'underline overline line-through'
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration-style').cssText is 'dashed'
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration-color').cssText is 'rgb(255, 0, 0)'
PASS checkComputedStyleValue() is true

Omitting text decoration line resets to its initial value "none" ("-webkit-text-decoration: navy dotted;"):
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration').cssText is 'none dotted rgb(0, 0, 128)'
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration-line').cssText is 'none'
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration-style').cssText is 'dotted'
PASS computedStyle.getPropertyCSSValue('-webkit-text-decoration-color').cssText is 'rgb(0, 0, 128)'
PASS checkComputedStyleValue() is true

PASS successfullyParsed is true

TEST COMPLETE
